Slight restructure pending graced tuplets
We use a PSGracedNote constructor and return a PSGracedNote in all
environments where grace notes are allowed (but usually, of course, grace
notes will be absent).  We no longer try to optimise by returning a
PSNote when there are no graces against the note.

This will later allow tuplets to hold graceable notes.